Scope fix for enforce_hyw_treaty_effect
FROM scope doesn't work here because the effect is called in a decision. Inside the decision however the code is pasted directly by the parser, so the scope we want to check is simply who is clicking the decision. This is the ROOT scope, hence no need to define the scope