HYW Merge Request
Partial Merge with the dev branch.
Some of my flavour rework is untested so I decided to only merge those that I have tried out myself. This should make the HYW a bit less boring and prolonged.
All the past and relevant changes are in changelog. (Those that didn't make into this branch weren't included in the changelog)