Rework Low Countries province-level privileges
I've removed events that were not called, replaced them with a decision opening an event where you can specify which privilege you are revoking. Costs are missing, but they were not present in the previous version as well, will need to be balanced together with the whole chain. It saved >2000 lines of code.